resolves #523, cleaned up sht31 local detect and refactored utilities module into env module
