+++ thermostatsupervisor/utilities.py
@@ -4,23 +4,12 @@
 import argparse
 import configparser
 import datetime
-import importlib.util
 import inspect
 import os
 import platform
 import socket
 import sys
-import traceback
 
-# third party libraries
-import psutil
-
-# thermostat config files
-from thermostatsupervisor import honeywell_config
-from thermostatsupervisor import kumocloud_config
-from thermostatsupervisor import kumolocal_config
-from thermostatsupervisor import mmm_config
-from thermostatsupervisor import sht31_config
 
 PACKAGE_NAME = 'thermostatsupervisor'  # should match name in __init__.py
 
@@ -47,96 +36,9 @@ DEBUG_LOG = 0x100  # print only if debug mode is on
 
 FILE_PATH = ".//data"
 MAX_LOG_SIZE_BYTES = 2**20  # logs rotate at this max size
-MIN_PYTHON_MAJOR_VERSION = 3  # minimum python major version required
-MIN_PYTHON_MINOR_VERSION = 7  # minimum python minor version required
-
-# all environment variables required by code should be registered here
-env_variables = {
-    "GMAIL_USERNAME": None,
-    "GMAIL_PASSWORD": None,
-}
-env_variables.update(honeywell_config.env_variables)
-env_variables.update(kumocloud_config.env_variables)
-env_variables.update(kumolocal_config.env_variables)
-env_variables.update(mmm_config.env_variables)
-env_variables.update(sht31_config.env_variables)
-
-
-def get_local_ip():
-    """Get local IP address for this PC."""
-    socket_obj = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    try:
-        # doesn't even have to be reachable
-        socket_obj.connect(('10.255.255.255', 1))
-        ip_address = socket_obj.getsockname()[0]
-    except Exception:
-        log_msg(traceback.format_exc(),
-                mode=BOTH_LOG, func_name=1)
-        ip_address = '127.0.0.1'
-    finally:
-        socket_obj.close()
-    return ip_address
-
 
 # set unit test IP address, same as client
 unit_test_mode = False  # in unit test mode
-unit_test_ip_address = get_local_ip()
-
-
-def get_env_variable(env_key):
-    """
-    Get environment variable.
-
-    Results will be logged but passwords will be masked off.
-
-    inputs:
-       env_key(str): env variable of interest
-       debug(bool): verbose debugging
-    returns:
-       (dict): {status, value, key}
-    """
-    # defaults
-    return_buffer = {
-        "status": NO_ERROR,
-        "value": None,
-        "key": env_key,
-    }
-
-    try:
-        # unit test key is not required to be in env var list
-        if env_key == sht31_config.UNIT_TEST_ENV_KEY:
-            return_buffer["value"] = unit_test_ip_address
-        else:
-            return_buffer["value"] = os.environ[env_key]
-
-        # mask off any password keys
-        if "PASSWORD" in return_buffer["key"]:
-            value_shown = "(hidden)"
-        else:
-            value_shown = return_buffer["value"]
-
-        log_msg(f"{env_key}={value_shown}",
-                mode=DEBUG_LOG)
-    except KeyError:
-        log_msg(f"FATAL ERROR: required environment variable '{env_key}'"
-                " is missing.", mode=CONSOLE_LOG + DATA_LOG)
-        return_buffer["status"] = ENVIRONMENT_ERROR
-    return return_buffer
-
-
-def load_all_env_variables():
-    """
-    Load all environment variables into a dictionary.
-
-    inputs:
-        None
-    returns:
-        None, populates env_variables dict.
-    """
-    for key in env_variables:
-        log_msg(f"checking key: {key}",
-                mode=BOTH_LOG, func_name=1)
-        env_variables[key] = get_env_variable(key)["value"]
 
 
 def get_function_name(stack_value=1):
@@ -297,20 +199,6 @@ def utf8len(input_string):
     return len(input_string.encode('utf-8'))
 
 
-def is_interactive_environment():
-    """Return True if script is run through IDE."""
-    parent = psutil.Process(os.getpid()).parent().name()
-    if parent == "cmd.exe":
-        return False
-    elif parent == "bash":
-        return False
-    elif parent == "eclipse.exe":
-        return True
-    else:
-        print(f"DEBUG: parent process={parent}")
-        raise OSError(f"unrecognized environment: {parent}")
-
-
 def temp_value_with_units(raw, disp_unit='F', precision=1) -> str:
     """
     Return string representing temperature and units.
@@ -430,11 +318,9 @@ def f_to_c(tempf) -> float:
 def is_host_on_local_net(host_name, ip_address=None):
     """
     Return True if specified host is on local network.
-
     socket.gethostbyaddr() throws exception for some IP address
     so preferred way to use this function is to pass in only the
     hostname and leave the IP as default (None).
-
     inputs:
         host_name(str): expected host name.
         ip_address(str): target IP address on local net.
@@ -467,106 +353,6 @@ def is_host_on_local_net(host_name, ip_address=None):
             print(f"DEBUG: expected host={host_name}, "
                   f"actual host={host_found}")
             return False, None
-
-
-def is_azure_environment():
-    """
-    Return True if machine is Azure pipeline.
-
-    Function assumes '192.' IP addresses are not Azure,
-    everything else is Azure.
-    """
-    return '192.' not in get_local_ip()
-
-
-def get_python_version(min_major_version=MIN_PYTHON_MAJOR_VERSION,
-                       min_minor_version=MIN_PYTHON_MINOR_VERSION,
-                       display_version=True):
-    """
-    Print current Python version to the screen.
-
-    inputs:
-        min_major_version(int): min allowed major version
-        min_minor_version(int): min allowed minor version
-        display_version(bool): True to print to screen.
-    return:
-        (tuple): (major version, minor version)
-    """
-    major_version = sys.version_info.major
-    minor_version = sys.version_info.minor
-
-    # display version
-    if display_version:
-        print(f"running on Python version {major_version}.{minor_version}")
-
-    # check major version
-    major_version_fail = False
-    if min_major_version is not None:
-        if not isinstance(min_major_version, (int, float)):
-            raise TypeError(f"input parameter 'min_major_version is type "
-                            f"({type(min_major_version)}), not int or float")
-        if major_version < min_major_version:
-            major_version_fail = True
-
-    # check major version
-    minor_version_fail = False
-    if min_minor_version is not None:
-        if not isinstance(min_minor_version, (int, float)):
-            raise TypeError(f"input parameter 'min_minor_version is type "
-                            f"({type(min_minor_version)}), not int or float")
-        if minor_version < min_minor_version:
-            minor_version_fail = True
-
-    if major_version_fail or minor_version_fail:
-        raise OSError(
-            f"current python major version ({major_version}.{minor_version}) "
-            f"is less than min python version required "
-            f"({min_major_version}.{min_minor_version})")
-
-    return (major_version, minor_version)
-
-
-def dynamic_module_import(name, path=None):
-    """
-    Find and load python module.
-
-    TODO: this module results in a resourcewarning within unittest:
-    sys:1: ResourceWarning: unclosed <socket.socket fd=628,
-    family=AddressFamily.AF_INET, type=SocketKind.SOCK_DGRAM, proto=0,
-    laddr=('0.0.0.0', 64963)>
-
-    inputs:
-        name(str): module name
-        path(str): file path (either relative or abs path),
-                   if path is None then will import from installed packages
-    returns:
-        mod(module): module object
-    """
-    try:
-        if path:
-            # local file import from relative or abs path
-            print(f"DEBUG attempting local import from {os.getcwd()}...")
-            print(f"target dir contents={os.listdir(path)}")
-            sys.path.insert(1, path)
-            mod = importlib.import_module(name)
-            if mod is None:
-                raise ModuleNotFoundError(f"module '{name}' could not "
-                                          f"be found at {path}")
-        else:
-            # installed package import
-            spec = importlib.util.find_spec(name, path)
-            if spec is None:
-                raise ModuleNotFoundError(f"module '{name}' could "
-                                          "not be found")
-            mod = spec.loader.load_module()
-    except Exception as ex:
-        log_msg(traceback.format_exc(),
-                mode=BOTH_LOG, func_name=1)
-        log_msg("module load failed: " + name,
-                mode=BOTH_LOG, func_name=1)
-        raise ex
-    else:
-        return mod
 
 
 # default parent_key if user_inputs are not pulled from file
